requirements.txt
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.0
shap==0.44.1
openai==1.12.0
tqdm==4.66.4
pyyaml==6.0.1
matplotlib==3.9.1
seaborn==0.13.2
jupyterlab==4.2.1
mlflow==2.13.1
python-dotenv==1.0.1
